requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
importlib-metadata==1.7.0
kombu==4.6.11
paho-mqtt==1.5.0
python-dotenv==0.14.0
pytz==2020.1
redis==3.5.3
vine==1.3.0
zipp==3.1.0
